Move the post-processor header stuctures into directoy, too
