+++ link-grammar/api.c
@@ -27,6 +27,7 @@
 #include "fast-match.h"
 #include "linkage.h"
 #include "post-process/post-process.h"
+#include "post-process/pp-structures.h"
 #include "preparation.h"
 #include "print/print.h"
 #include "prune.h"
