+++ link-grammar/constituents.c
@@ -18,6 +18,7 @@
 #include "error.h"
 #include "externs.h"
 #include "post-process/post-process.h"
+#include "post-process/pp-structures.h"
 #include "print/print-util.h"
 #include "string-set.h"
