+++ link-grammar/post-process/pp_knowledge.c
@@ -23,6 +23,7 @@
 #include "pp_lexer.h"
 #include "pp_linkset.h"
 #include "string-set.h"
+#include "utilities.h"
 
 #define D_PPK 10                       /* verbosity level for this file */
 #define PP_MAX_UNIQUE_LINK_NAMES 1024  /* just needs to be approximate */
