+++ link-grammar/prune.c
@@ -17,6 +17,7 @@
 #include "disjunct-utils.h"
 #include "externs.h"
 #include "post-process/post-process.h"
+#include "post-process/pp-structures.h"
 #include "print/print.h"
 
 #include "prune.h"
